flask_api/requirements.txt
flask 
flask-cors 
pillow
numpy==1.26.2
scikit-learn==1.3.2
Cython==3.0.6
transformers>=4.35.2
opencv-python==4.8.1.78
matplotlib 
tqdm
diffusers[torch]==0.24.0
accelerate==0.25.0
git+https://github.com/openai/CLIP.git
controlnet-aux==0.0.7